pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "langfuse>=3.0.9",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
]
